refactor: route src.retrieval imports through public API
Move all cross-package imports of retrieval internals to use
src.retrieval public exports instead of reaching into internal
modules. Adds warm_up_ollama and call_ollama to the retrieval
package API and updates 7 caller files (server/worker,
server/activities, and 4 guardrails modules).
