+++ src/retrieval/__init__.py
@@ -1,6 +1,6 @@
 # @summary
 # Retrieval package public API.
-# Exports: RAGRequest, RAGChain, RAGResponse, process_query, QueryResult, QueryAction, LocalBGEReranker, RankedResult, OllamaGenerator, ConfidenceBreakdown, PostGuardrailAction, compute_composite_confidence
+# Exports: RAGRequest, RAGChain, RAGResponse, process_query, QueryResult, QueryAction, LocalBGEReranker, RankedResult, OllamaGenerator, ConfidenceBreakdown, PostGuardrailAction, compute_composite_confidence, warm_up_ollama, call_ollama
 # Deps: src.retrieval.common, src.retrieval.pipeline, src.retrieval.query, src.retrieval.generation
 # @end-summary
 
@@ -12,9 +12,11 @@ from src.retrieval.common import (
 from src.retrieval.query import (
     QueryAction,
     QueryResult,
+    warm_up_ollama,
 )
 from src.retrieval.query.nodes import process_query
 from src.retrieval.query.nodes import LocalBGEReranker
+from src.retrieval.query.nodes import _call_ollama as call_ollama
 from src.retrieval.pipeline import RAGChain
 from src.retrieval.generation.nodes import OllamaGenerator
 from src.retrieval.generation.confidence import (
@@ -36,4 +38,6 @@ __all__ = [
     "ConfidenceBreakdown",
     "PostGuardrailAction",
     "compute_composite_confidence",
+    "warm_up_ollama",
+    "call_ollama",
 ]
